added new function checkAuthentication
